src/agent/requirements.txt
azure-ai-projects>=1.0.0b9
azure-identity>=1.17.0
azure-search-documents>=11.5.0
openai>=1.40.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.30.0
